requirements.txt
# Web Automation
selenium>=4.15.0
webdriver-manager>=4.0.0

# Data Processing
pandas>=2.1.0
openpyxl>=3.1.2

# Phone Number Validation
phonenumbers>=8.13.0

# Environment Configuration
python-dotenv>=1.0.0

# HTTP and URL handling
urllib3>=2.0.0

# Testing (optional, for development)
pytest>=7.4.0
pytest-cov>=4.1.0

# Code Quality (optional, for development)
black>=23.0.0
flake8>=6.0.0
pylint>=3.0.0
